Completed tasks on tree(balande, perfect, uncle)

+++ 14-binary_tree_balance.c
@@ -7,9 +7,9 @@
  * Return: 0 if tree is NULL
  */
 
-size_t binary_tree_height(const binary_tree_t *tree)
+int binary_tree_height(const binary_tree_t *tree)
 {
-	size_t left_tree, right_tree;
+	int left_tree, right_tree;
 
 	if (tree == NULL)
 		return (0);
